requirements.txt
py-cord
datetime
python_dotenv
requests